enhancement:  modular billing API architecture and comprehensive testing
